requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-restful==0.3.10
flask-cors==4.0.0
flask-limiter==3.5.0
flask-mail==0.9.1
flask-socketio==5.3.5
python-socketio==5.10.0
requests==2.31.0
werkzeug==3.0.1
python-dotenv==1.0.0
pysocks==1.7.1
reportlab==4.0.7
openpyxl==3.1.2
pytest==7.4.3
pytest-cov==4.1.0
hypothesis==6.88.1